fix preferences panel and add buttons to install and uninstall template
