+++ molecularnodes/ui/pref.py
@@ -1,18 +1,30 @@
 import bpy
 from bpy.types import AddonPreferences
 
-from ..utils import template_install
+from ..template import template_install, template_uninstall
+from .. import __package__
 
-install_instructions = "https://bradyajohnston.github.io/MolecularNodes/installation.html#installing-biotite-mdanalysis"
 
-
-class MN_OT_Install_Template(bpy.types.Operator):
-    bl_idname = "mn.install_template"
+class MN_OT_Template_Install(bpy.types.Operator):
+    bl_idname = "mn.template_install"
     bl_label = "Install Template File"
     bl_description = "Install the Molecular Nodes startup template file."
 
-    def exectute(self, context):
+    def execute(self, context):
         template_install()
+        self.report({"INFO"}, "Installed Molecular Nodes template.")
+        return {"FINISHED"}
+
+
+class MN_OT_Template_Uninstall(bpy.types.Operator):
+    bl_idname = "mn.template_uninstall"
+    bl_label = "Uninstall Template"
+    bl_description = "Uninstall the Molecular Nodes startup template file."
+
+    def execute(self, context):
+        template_uninstall()
+        self.report({"INFO"}, "Uninstalled Molecular Nodes template.")
+        return {"FINISHED"}
 
 
 class MolecularNodesPreferences(AddonPreferences):
@@ -20,8 +32,9 @@ class MolecularNodesPreferences(AddonPreferences):
 
     def draw(self, context):
         layout = self.layout
-        layout.label(text="testing")
-        layout.operator("mn.install_template", text="Install Template")
+        row = layout.row()
+        row.operator("mn.template_install")
+        row.operator("mn.template_uninstall")
 
 
-CLASSES = [MN_OT_Install_Template, MolecularNodesPreferences]
+CLASSES = [MN_OT_Template_Install, MN_OT_Template_Uninstall, MolecularNodesPreferences]
